+++ R/print-all.R
@@ -6,7 +6,13 @@
 #' @param x An object used to select a method.
 #' @examples
 #' print_all(state.name)
+#' @importFrom tibble is_tibble
 #' @export
 print_all <- function(x) {
-  print(x, n = Inf)
+  if (tibble::is_tibble(x)) {
+    print(x, n = Inf)
+  } else {
+    print(x)
+    warning("intended for use with tibble rows")
+  }
 }