+++ R/str-dist.R
@@ -5,7 +5,7 @@
 #' @param a `R` object (target); will be converted by [base::as.character()].
 #' @param b `R` object (source); will be converted by [base::as.character()].
 #' @param method Method for distance calculation. The default is "osa."
-#' @param ... Other rguments passed to [stringdist::stringdist()].
+#' @param ... Other arguments passed to [stringdist::stringdist()].
 #' @return The distance between string `a` and string `b`.
 #' @importFrom stringdist stringdist
 #' @examples