+++ data-raw/data_usps.R
@@ -8,24 +8,20 @@ library(rvest)
 
 # Publication 28 - Postal Addressing Standards
 # Appendix B - Two–Letter State and Possession Abbreviations
-usps_b0 <- "https://pe.usps.com/text/pub28/28apb.htm"
+usps_b0 <- read_html("https://pe.usps.com/text/pub28/28apb.htm")
 
 # scrape states table
-usps_state <-
-  read_html(x = usps_b0) %>%
+usps_state <- usps_b0 %>%
   html_node(css = "#ep18684") %>%
   html_table(fill = TRUE, header = TRUE) %>%
   set_names(nm = c("full", "abb")) %>%
-  select(abb, full) %>%
   mutate(full = str_to_upper(full))
 
 # scrape armed forces table
-usps_armed <-
-  read_html(x = usps_b0) %>%
+usps_armed <- usps_b0 %>%
   html_node(css = "#ep19241") %>%
   html_table(fill = TRUE, header = TRUE) %>%
   set_names(nm = c("full", "abb")) %>%
-  select(abb, full) %>%
   mutate(full = str_to_upper(str_remove(full, "([:punct:].*)$")))
 
 # combined tables
@@ -38,7 +34,7 @@ usps_state <- usps_state %>%
 usethis::use_data(usps_state, overwrite = TRUE)
 write_lines(
   x = usps_state,
-  path = "data-raw/.csv"
+  path = "data-raw/usps_state.csv"
 )
 
 # save state names vector
@@ -61,34 +57,31 @@ write_lines(
 
 # Publication 28 - Postal Addressing Standards
 # Appendix C - C1 Street Suffix Abbreviations
-usps_c1 <- "https://pe.usps.com/text/pub28/28apc_002.htm"
+usps_c1 <- read_html("https://pe.usps.com/text/pub28/28apc_002.htm")
 
 # scrape states table
-usps_street <-
-  read_html(x = usps_c1) %>%
+usps_street <- usps_c1 %>%
   html_node(css = "#ep533076") %>%
   html_table(fill = TRUE, header = TRUE) %>%
-  select(1, 2) %>%
+  select(2, 3) %>%
   set_names(nm = c("full", "abb")) %>%
-  select(abb, full) %>%
   filter(full != abb) %>%
   as_tibble()
 
 # Publication 28 - Postal Addressing Standards
 # Appendix C - C2 Secondary Unit Designators
-usps_c2 <- "https://pe.usps.com/text/pub28/28apc_003.htm"
+usps_c2 <- read_html("https://pe.usps.com/text/pub28/28apc_003.htm")
 
 # scrape states table
-usps_unit <-
-  read_html(x = usps_c2) %>%
+usps_unit <- usps_c2 %>%
   html_node(css = "#ep538257") %>%
   html_table(fill = TRUE, header = TRUE) %>%
   as_tibble(.name_repair = "unique") %>%
   slice(-26) %>%
-  select(3, 1) %>%
+  select(1, 3) %>%
   na_if("") %>%
   drop_na() %>%
-  set_names(nm = c("abb", "full")) %>%
+  set_names(nm = c("full", "abb")) %>%
   mutate(
     full = str_to_upper(full) %>% str_remove_all("[^\\w]"),
     abb = str_remove_all(abb, "[^\\w]")
@@ -96,12 +89,10 @@ usps_unit <-
   filter(full != abb)
 
 # scrape geographic directions
-usps_dirs <-
-  read_html(x = usps_b0) %>%
+usps_dirs <- usps_b0 %>%
   html_node(css = "#ep19168") %>%
   html_table(fill = TRUE, header = TRUE) %>%
   set_names(nm = c("full", "abb")) %>%
-  select(abb, full) %>%
   mutate(full = str_to_upper(full))
 
 # combine tables
@@ -121,6 +112,7 @@ write_lines(
 
 # create a subet of abbs in city names
 usps_city <- usps_street %>%
+  select(2, 1) %>%
   filter(
     # keep only those used
     full %in% valid_city,
@@ -129,8 +121,8 @@ usps_city <- usps_street %>%
   ) %>%
   # add those not included
   add_row(
-    full = c("FORT", "HEIGHTS", "MOUNT", "MOUNTAIN", "PORT", "SAINT", "TOWNSHIP"),
-    abb  = c("FT",   "HTS",     "MT",    "MTN",      "PRT",  "ST",    "TWP")
+    abb  = c("FT",   "HTS",     "MT",    "MTN",      "PRT",  "ST",    "TWP"),
+    full = c("FORT", "HEIGHTS", "MOUNT", "MOUNTAIN", "PORT", "SAINT", "TOWNSHIP")
   ) %>%
   # add directional abbs
   bind_rows(usps_dirs) %>%