+++ NEWS.md
@@ -1,3 +1,15 @@
+# campfin 0.2.2
+
+* `flag_dupes()` now also calls `duplicated(fromLast = TRUE)` to capture all.
+* Separate address `[:digit:]` from `[:alpha:]` with space.
+* Create `prop_distinct()`.
+
+# campfin 0.2.1
+
+* The `normal_address()` now calls `abbrev_full()` instead of `expand_abbrev()`.
+* Reverse order of columns in `usps_street`, `usps_state`, and `usps_city`.
+* Update to version 2.0 of the `CODE_OF_CONDUCT.md` and URL.
+
 # campfin 0.1.1.9000
 
 * Added a `NEWS.md` file to track changes to the package.